poc/reid_module/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.116.0

fsspec==2025.5.1
gdown==5.2.0

grpcio==1.73.1
h11==0.16.0
httptools==0.6.4
idna==3.10

Markdown==3.8.2

mkl-service==2.4.0

msgpack==1.1.1

numpy==2.2.6
opencv-python==4.12.0.88

protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
PySocks==1.7.1
python-dotenv==1.1.1

pyzmq==27.0.0
requests==2.32.4
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.5.1
torchreid==0.2.5

tqdm==4.67.1
typing-inspection==0.4.1

urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
